Allow name to be optional for create command
Currently, the create command requires a name to be provided,
and this is used for a suffix on the filename and then as the
class name. This PR makes the name optional, such that if
it is omitted, the system will automatically generate a name
that is valid. The filename is "<datetime>_migration.php" and
the class name is "V<Datetime>Migration".

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>