better handle when schema included on second table for postgres rename
Currently, if one were to do `renameTable('schema1.table', 'schema2.table')`, phinx
under the hood will translate that into a rename of "schema1"."table" to
"schema1"."schema2.table". This is almost certainly not what someone might want
given how schemas are handled in other commands (and the first argument here).
However, it is also illegal to try and rename a column from one schema into another
using the ALTER TABLE _ RENAME SQL (it requires a different SQL statement structure).
As such, this PR makes it so that if someone attempts to rename a table where
the second argument contains a schema (like the above example), it'll parse
that into composite parts and check if the schema of the second table matches
that of the first table, and if so, silently fixes things so the rename will
work on just the table name, but if they're different, will issue an error.

For example, doing:

renameTable("schema1.table", "schema1.table2") -> "schema1"."table" RENAME TO "table2"
renameTable("schema1.table", "schema2.table2") -> throw \InvalidArgumentException

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>